Add map basemaps, distance calc, and filters
Change Road.DistanceFeet from double? to decimal? and update Roads UI to bind TValue accordingly. Add proj4.js dependency and enhance road location editor and client JS: support basemap selector, service-area GeoJSON loading, distance calculation (projecting to NAD83 Alaska SP zone 3 via proj4, converting meters to US survey feet), existing-line distance reporting, and apply-calculated-distance integration with the Roads editor. Introduce CustomQuickFilter support in EntityManagementLayout and add quick filters for Ordinances and Roads. Add service-area mileage calculation (sums linked roads' DistanceFeet and converts to miles) in ServiceAreas page. Expose feature layer URL/query parameters from Roads to the map editor; use configuration key FeatureLayer:ServiceAreasUrl. Notes: JS changes are in wwwroot/js/roadLocationMap.js and App.razor now loads proj4; these changes may require a DB migration or review for the type change of DistanceFeet.

+++ rs-ruralia.Shared/Models/Road.cs
@@ -17,7 +17,7 @@ public class Road
 
   [Column("distance_feet")]
   [Display(Name = "Distance (Feet)", Description = "Distance used for calculating quotes.")]
-  public double? DistanceFeet { get; set; }
+  public decimal? DistanceFeet { get; set; }
 
   [Column("start_description")]
   [MaxLength(250)]
